+++ spec/infra/harness/variables.tf
@@ -9,8 +9,6 @@ variable "component" {}
 variable "deployment_identifier" {}
 variable "dependencies" {}
 
-variable "domain_name" {}
-variable "public_zone_id" {}
 variable "private_zone_id" {}
 
 variable "infrastructure_events_bucket" {}