+++ spec/support/shared_contexts/terraform.rb
@@ -5,14 +5,15 @@ require_relative '../terraform_module'
 shared_context :terraform do
   include Awspec::Helper::Finder
 
-  let(:vars) { TerraformModule.configuration.vars}
+  let(:vars) {TerraformModule.configuration.for(:harness).vars}
 
-  def output_with_name(name)
-    TerraformModule.output_with_name(name)
+  def output_for(role, name)
+    TerraformModule.output_for(role, name)
   end
 
   def reprovision(override_vars)
     TerraformModule.provision(
-        TerraformModule.configuration.vars.to_h.merge(override_vars))
+        TerraformModule.configuration.for(:harness)
+            .vars.to_h.merge(override_vars))
   end
 end